+++ src/main/java/com/rbkmoney/deanonimus/db/exception/ContractNotFoundException.java
@@ -1,4 +1,9 @@
 package com.rbkmoney.deanonimus.db.exception;
 
 public class ContractNotFoundException extends RuntimeException {
+
+    public ContractNotFoundException(String id) {
+        super("Not found with id: " + id);
+    }
+
 }
