+++ src/main/java/com/rbkmoney/deanonimus/db/exception/ContractorNotFoundException.java
@@ -1,4 +1,9 @@
 package com.rbkmoney.deanonimus.db.exception;
 
 public class ContractorNotFoundException extends RuntimeException {
+
+    public ContractorNotFoundException(String id) {
+        super("Not found with id: " + id);
+    }
+
 }
