+++ src/main/java/com/rbkmoney/deanonimus/db/exception/PartyNotFoundException.java
@@ -1,4 +1,8 @@
 package com.rbkmoney.deanonimus.db.exception;
 
 public class PartyNotFoundException extends RuntimeException {
+
+    public PartyNotFoundException(String id) {
+        super("Not found with id: " + id);
+    }
 }
