+++ src/main/java/com/rbkmoney/deanonimus/db/exception/ShopNotFoundException.java
@@ -1,4 +1,8 @@
 package com.rbkmoney.deanonimus.db.exception;
 
 public class ShopNotFoundException extends RuntimeException {
+
+    public ShopNotFoundException(String id) {
+        super("Not found with id: " + id);
+    }
 }
