+++ src/main/java/com/rbkmoney/deanonimus/kafka/handler/PartyManagementHandlerService.java
@@ -2,8 +2,7 @@ package com.rbkmoney.deanonimus.kafka.handler;
 
 import com.rbkmoney.damsel.payment_processing.EventPayload;
 import com.rbkmoney.damsel.payment_processing.PartyChange;
-import com.rbkmoney.damsel.payment_processing.PartyEventData;
-import com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.PartyManagementHandler;
+import com.rbkmoney.deanonimus.kafka.handler.party_management.PartyManagementHandler;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
 import com.rbkmoney.sink.common.parser.impl.MachineEventParser;
 import lombok.RequiredArgsConstructor;
@@ -20,8 +19,7 @@ import java.util.List;
 public class PartyManagementHandlerService {
 
     private final List<PartyManagementHandler> partyManagementHandlers;
-    private final MachineEventParser<PartyEventData> parser;
-    private final MachineEventParser<EventPayload> parser2;
+    private final MachineEventParser<EventPayload> parser;
 
     @Transactional(propagation = Propagation.REQUIRED)
     public void handleEvents(List<MachineEvent> machineEvents) {
@@ -29,7 +27,7 @@ public class PartyManagementHandlerService {
     }
 
     private void handleIfAccept(MachineEvent machineEvent) {
-        EventPayload eventPayload = parser2.parse(machineEvent);
+        EventPayload eventPayload = parser.parse(machineEvent);
         if (eventPayload.isSetPartyChanges()) {
             final List<PartyChange> partyChanges = eventPayload.getPartyChanges();
             for (int i = 0; i < partyChanges.size(); i++) {
