+++ src/main/java/com/rbkmoney/deanonimus/kafka/handler/party_management/AbstractClaimChangedHandler.java
@@ -1,4 +1,4 @@
-package com.rbkmoney.deanonimus.kafka.handler.party_mngmnt;
+package com.rbkmoney.deanonimus.kafka.handler.party_management;
 
 import com.rbkmoney.damsel.payment_processing.ClaimStatus;
 import com.rbkmoney.damsel.payment_processing.PartyChange;
