+++ src/main/java/com/rbkmoney/deanonimus/kafka/handler/party_management/PartyManagementHandler.java
@@ -1,4 +1,4 @@
-package com.rbkmoney.deanonimus.kafka.handler.party_mngmnt;
+package com.rbkmoney.deanonimus.kafka.handler.party_management;
 
 import com.rbkmoney.damsel.payment_processing.PartyChange;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
