+++ src/main/java/com/rbkmoney/deanonimus/kafka/handler/party_management/contract/ContractLegalAgreementBoundHandler.java
@@ -1,6 +1,6 @@
-package com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.contract;
+package com.rbkmoney.deanonimus.kafka.handler.party_management.contract;
 
-import com.rbkmoney.damsel.domain.ReportPreferences;
+import com.rbkmoney.damsel.domain.LegalAgreement;
 import com.rbkmoney.damsel.payment_processing.ClaimEffect;
 import com.rbkmoney.damsel.payment_processing.ContractEffectUnit;
 import com.rbkmoney.damsel.payment_processing.PartyChange;
@@ -9,8 +9,7 @@ import com.rbkmoney.deanonimus.db.exception.ContractNotFoundException;
 import com.rbkmoney.deanonimus.db.exception.PartyNotFoundException;
 import com.rbkmoney.deanonimus.domain.Contract;
 import com.rbkmoney.deanonimus.domain.Party;
-import com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.AbstractClaimChangedHandler;
-import com.rbkmoney.deanonimus.util.ContractUtil;
+import com.rbkmoney.deanonimus.kafka.handler.party_management.AbstractClaimChangedHandler;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,7 +22,7 @@ import java.util.List;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ContractReportPreferencesChangedHandler extends AbstractClaimChangedHandler {
+public class ContractLegalAgreementBoundHandler extends AbstractClaimChangedHandler {
 
     private final PartyRepository partyRepository;
 
@@ -33,7 +32,7 @@ public class ContractReportPreferencesChangedHandler extends AbstractClaimChange
         long sequenceId = event.getEventId();
         List<ClaimEffect> claimEffects = getClaimStatus(change).getAccepted().getEffects();
         for (ClaimEffect claimEffect : claimEffects) {
-            if (claimEffect.isSetContractEffect() && claimEffect.getContractEffect().getEffect().isSetReportPreferencesChanged()) {
+            if (claimEffect.isSetContractEffect() && claimEffect.getContractEffect().getEffect().isSetLegalAgreementBound()) {
                 handleEvent(event, changeId, sequenceId, claimEffect);
             }
         }
@@ -41,25 +40,20 @@ public class ContractReportPreferencesChangedHandler extends AbstractClaimChange
 
     private void handleEvent(MachineEvent event, Integer changeId, long sequenceId, ClaimEffect claimEffect) {
         ContractEffectUnit contractEffectUnit = claimEffect.getContractEffect();
-        ReportPreferences reportPreferencesChanged = contractEffectUnit.getEffect().getReportPreferencesChanged();
+        LegalAgreement legalAgreement = contractEffectUnit.getEffect().getLegalAgreementBound();
         String contractId = contractEffectUnit.getContractId();
         String partyId = event.getSourceId();
-        log.info("Start contract report preferences changed handling, sequenceId={}, partyId={}, contractId={}, changeId={}",
+        log.info("Start contract legal agreement bound handling, sequenceId={}, partyId={}, contractId={}, changeId={}",
                 sequenceId, partyId, contractId, changeId);
 
-        Party party = partyRepository.findById(partyId).orElseThrow(PartyNotFoundException::new);
+        Party party = partyRepository.findById(partyId).orElseThrow(() -> new PartyNotFoundException(partyId));
 
-        Contract contract = party.getContractById(contractId).orElseThrow(ContractNotFoundException::new);
-
-        if (reportPreferencesChanged != null && reportPreferencesChanged.isSetServiceAcceptanceActPreferences()) {
-            ContractUtil.fillReportPreferences(contract, reportPreferencesChanged.getServiceAcceptanceActPreferences());
-        } else {
-            ContractUtil.setNullReportPreferences(contract);
-        }
+        Contract contract = party.getContractById(contractId).orElseThrow(() -> new ContractNotFoundException(contractId));
+        contract.setLegalAgreementId(legalAgreement.getLegalAgreementId());
 
         partyRepository.save(party);
 
-        log.info("End contract report preferences changed handling, sequenceId={}, partyId={}, contractId={}, changeId={}",
+        log.info("End contract legal agreement bound handling, sequenceId={}, partyId={}, contractId={}, changeId={}",
                 sequenceId, partyId, contractId, changeId);
     }
 }
