+++ src/main/java/com/rbkmoney/deanonimus/kafka/handler/party_management/contract/ContractReportPreferencesChangedHandler.java
@@ -1,6 +1,6 @@
-package com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.contract;
+package com.rbkmoney.deanonimus.kafka.handler.party_management.contract;
 
-import com.rbkmoney.damsel.domain.LegalAgreement;
+import com.rbkmoney.damsel.domain.ReportPreferences;
 import com.rbkmoney.damsel.payment_processing.ClaimEffect;
 import com.rbkmoney.damsel.payment_processing.ContractEffectUnit;
 import com.rbkmoney.damsel.payment_processing.PartyChange;
@@ -9,7 +9,8 @@ import com.rbkmoney.deanonimus.db.exception.ContractNotFoundException;
 import com.rbkmoney.deanonimus.db.exception.PartyNotFoundException;
 import com.rbkmoney.deanonimus.domain.Contract;
 import com.rbkmoney.deanonimus.domain.Party;
-import com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.AbstractClaimChangedHandler;
+import com.rbkmoney.deanonimus.kafka.handler.party_management.AbstractClaimChangedHandler;
+import com.rbkmoney.deanonimus.util.ContractUtil;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,7 +23,7 @@ import java.util.List;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ContractLegalAgreementBoundHandler extends AbstractClaimChangedHandler {
+public class ContractReportPreferencesChangedHandler extends AbstractClaimChangedHandler {
 
     private final PartyRepository partyRepository;
 
@@ -31,30 +32,34 @@ public class ContractLegalAgreementBoundHandler extends AbstractClaimChangedHand
     public void handle(PartyChange change, MachineEvent event, Integer changeId) {
         long sequenceId = event.getEventId();
         List<ClaimEffect> claimEffects = getClaimStatus(change).getAccepted().getEffects();
-        for (int i = 0; i < claimEffects.size(); i++) {
-            ClaimEffect claimEffect = claimEffects.get(i);
-            if (claimEffect.isSetContractEffect() && claimEffect.getContractEffect().getEffect().isSetLegalAgreementBound()) {
-                handleEvent(event, changeId, sequenceId, claimEffects.get(i));
+        for (ClaimEffect claimEffect : claimEffects) {
+            if (claimEffect.isSetContractEffect() && claimEffect.getContractEffect().getEffect().isSetReportPreferencesChanged()) {
+                handleEvent(event, changeId, sequenceId, claimEffect);
             }
         }
     }
 
     private void handleEvent(MachineEvent event, Integer changeId, long sequenceId, ClaimEffect claimEffect) {
         ContractEffectUnit contractEffectUnit = claimEffect.getContractEffect();
-        LegalAgreement legalAgreement = contractEffectUnit.getEffect().getLegalAgreementBound();
+        ReportPreferences reportPreferencesChanged = contractEffectUnit.getEffect().getReportPreferencesChanged();
         String contractId = contractEffectUnit.getContractId();
         String partyId = event.getSourceId();
-        log.info("Start contract legal agreement bound handling, sequenceId={}, partyId={}, contractId={}, changeId={}",
+        log.info("Start contract report preferences changed handling, sequenceId={}, partyId={}, contractId={}, changeId={}",
                 sequenceId, partyId, contractId, changeId);
 
-        Party party = partyRepository.findById(partyId).orElseThrow(PartyNotFoundException::new);
+        Party party = partyRepository.findById(partyId).orElseThrow(() -> new PartyNotFoundException(partyId));
 
-        Contract contract = party.getContractById(contractId).orElseThrow(ContractNotFoundException::new);
-        contract.setLegalAgreementId(legalAgreement.getLegalAgreementId());
+        Contract contract = party.getContractById(contractId).orElseThrow(() -> new ContractNotFoundException(contractId));
+
+        if (reportPreferencesChanged != null && reportPreferencesChanged.isSetServiceAcceptanceActPreferences()) {
+            ContractUtil.fillReportPreferences(contract, reportPreferencesChanged.getServiceAcceptanceActPreferences());
+        } else {
+            ContractUtil.setNullReportPreferences(contract);
+        }
 
         partyRepository.save(party);
 
-        log.info("End contract legal agreement bound handling, sequenceId={}, partyId={}, contractId={}, changeId={}",
+        log.info("End contract report preferences changed handling, sequenceId={}, partyId={}, contractId={}, changeId={}",
                 sequenceId, partyId, contractId, changeId);
     }
 }
