+++ src/main/java/com/rbkmoney/deanonimus/kafka/handler/party_management/contractor/ContractorCreatedHandler.java
@@ -1,4 +1,4 @@
-package com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.contractor;
+package com.rbkmoney.deanonimus.kafka.handler.party_management.contractor;
 
 import com.rbkmoney.damsel.domain.PartyContractor;
 import com.rbkmoney.damsel.payment_processing.ClaimEffect;
@@ -7,23 +7,19 @@ import com.rbkmoney.damsel.payment_processing.PartyChange;
 import com.rbkmoney.deanonimus.db.PartyRepository;
 import com.rbkmoney.deanonimus.db.exception.PartyNotFoundException;
 import com.rbkmoney.deanonimus.domain.Party;
-import com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.AbstractClaimChangedHandler;
+import com.rbkmoney.deanonimus.kafka.handler.party_management.AbstractClaimChangedHandler;
 import com.rbkmoney.deanonimus.util.ContractorUtil;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-import static org.springframework.core.Ordered.HIGHEST_PRECEDENCE;
-
 @Slf4j
 @Component
-@Order(HIGHEST_PRECEDENCE)
 @RequiredArgsConstructor
 public class ContractorCreatedHandler extends AbstractClaimChangedHandler {
 
@@ -48,7 +44,7 @@ public class ContractorCreatedHandler extends AbstractClaimChangedHandler {
         String contractorId = contractorEffect.getId();
         String partyId = event.getSourceId();
         log.info("Start contractor created handling, eventId={}, partyId={}, contractorId={}", eventId, partyId, contractorId);
-        Party party = partyRepository.findById(partyId).orElseThrow(PartyNotFoundException::new);
+        Party party = partyRepository.findById(partyId).orElseThrow(() -> new PartyNotFoundException(partyId));
         com.rbkmoney.deanonimus.domain.Contractor contractor = ContractorUtil.convertContractor(partyId, contractorCreated, contractorId);
         contractor.setIdentificationalLevel(partyContractor.getStatus().name());
