+++ src/main/java/com/rbkmoney/deanonimus/kafka/handler/party_management/shop/ShopContractChangedHandler.java
@@ -1,14 +1,15 @@
-package com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.shop;
+package com.rbkmoney.deanonimus.kafka.handler.party_management.shop;
 
-import com.rbkmoney.damsel.domain.ShopAccount;
 import com.rbkmoney.damsel.payment_processing.ClaimEffect;
 import com.rbkmoney.damsel.payment_processing.PartyChange;
+import com.rbkmoney.damsel.payment_processing.ShopContractChanged;
 import com.rbkmoney.damsel.payment_processing.ShopEffectUnit;
 import com.rbkmoney.deanonimus.db.PartyRepository;
 import com.rbkmoney.deanonimus.db.exception.PartyNotFoundException;
+import com.rbkmoney.deanonimus.db.exception.ShopNotFoundException;
 import com.rbkmoney.deanonimus.domain.Party;
-import com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.AbstractClaimChangedHandler;
-import com.rbkmoney.deanonimus.util.ShopUtil;
+import com.rbkmoney.deanonimus.domain.Shop;
+import com.rbkmoney.deanonimus.kafka.handler.party_management.AbstractClaimChangedHandler;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,7 +22,7 @@ import java.util.List;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ShopAccountCreatedHandler extends AbstractClaimChangedHandler {
+public class ShopContractChangedHandler extends AbstractClaimChangedHandler {
 
     private final PartyRepository partyRepository;
 
@@ -31,28 +32,29 @@ public class ShopAccountCreatedHandler extends AbstractClaimChangedHandler {
         long sequenceId = event.getEventId();
         List<ClaimEffect> claimEffects = getClaimStatus(change).getAccepted().getEffects();
         for (ClaimEffect claimEffect : claimEffects) {
-            if (claimEffect.isSetShopEffect() && claimEffect.getShopEffect().getEffect().isSetAccountCreated()) {
+            if (claimEffect.isSetShopEffect() && claimEffect.getShopEffect().getEffect().isSetContractChanged()) {
                 handleEvent(event, changeId, sequenceId, claimEffect);
             }
         }
     }
 
-    private void handleEvent(MachineEvent event, Integer changeId, long sequenceId, ClaimEffect e) {
-        ShopEffectUnit shopEffect = e.getShopEffect();
-        ShopAccount accountCreated = shopEffect.getEffect().getAccountCreated();
+    private void handleEvent(MachineEvent event, Integer changeId, long sequenceId, ClaimEffect claimEffect) {
+        ShopEffectUnit shopEffect = claimEffect.getShopEffect();
+        ShopContractChanged contractChanged = shopEffect.getEffect().getContractChanged();
         String shopId = shopEffect.getShopId();
         String partyId = event.getSourceId();
-        log.info("Start shop accountCreated handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
+        log.info("Start shop contractChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
                 sequenceId, partyId, shopId, changeId);
 
-        Party party = partyRepository.findById(partyId).orElseThrow(PartyNotFoundException::new);
+        Party party = partyRepository.findById(partyId).orElseThrow(() -> new PartyNotFoundException(partyId));
+        Shop shop = party.getShopById(shopId).orElseThrow(() -> new ShopNotFoundException(shopId));
 
-        ShopUtil.fillShopAccount(party.getShops().get(0), accountCreated);
+        shop.setContractId(contractChanged.getContractId());
+        shop.setPayoutToolId(contractChanged.getPayoutToolId());
 
         partyRepository.save(party);
 
-        log.info("End shop accountCreated handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
+        log.info("End shop contractChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
                 sequenceId, partyId, shopId, changeId);
     }
-
 }
