+++ src/main/java/com/rbkmoney/deanonimus/kafka/handler/party_management/shop/ShopDetailsChangedHandler.java
@@ -1,14 +1,15 @@
-package com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.shop;
+package com.rbkmoney.deanonimus.kafka.handler.party_management.shop;
 
-import com.rbkmoney.damsel.domain.ShopLocation;
+import com.rbkmoney.damsel.domain.ShopDetails;
 import com.rbkmoney.damsel.payment_processing.ClaimEffect;
 import com.rbkmoney.damsel.payment_processing.PartyChange;
 import com.rbkmoney.damsel.payment_processing.ShopEffectUnit;
 import com.rbkmoney.deanonimus.db.PartyRepository;
 import com.rbkmoney.deanonimus.db.exception.PartyNotFoundException;
+import com.rbkmoney.deanonimus.db.exception.ShopNotFoundException;
 import com.rbkmoney.deanonimus.domain.Party;
 import com.rbkmoney.deanonimus.domain.Shop;
-import com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.AbstractClaimChangedHandler;
+import com.rbkmoney.deanonimus.kafka.handler.party_management.AbstractClaimChangedHandler;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,7 +22,7 @@ import java.util.List;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ShopLocationChangedHandler extends AbstractClaimChangedHandler {
+public class ShopDetailsChangedHandler extends AbstractClaimChangedHandler {
 
     private final PartyRepository partyRepository;
 
@@ -31,32 +32,30 @@ public class ShopLocationChangedHandler extends AbstractClaimChangedHandler {
         long sequenceId = event.getEventId();
         List<ClaimEffect> claimEffects = getClaimStatus(change).getAccepted().getEffects();
         for (ClaimEffect claimEffect : claimEffects) {
-            if (claimEffect.isSetShopEffect() && claimEffect.getShopEffect().getEffect().isSetLocationChanged()) {
+            if (claimEffect.isSetShopEffect() && claimEffect.getShopEffect().getEffect().isSetDetailsChanged()) {
                 handleEvent(event, changeId, sequenceId, claimEffect);
             }
         }
     }
 
-    private void handleEvent(MachineEvent event, Integer changeId, long sequenceId, ClaimEffect claimEffect) {
-        ShopEffectUnit shopEffect = claimEffect.getShopEffect();
-        ShopLocation locationChanged = shopEffect.getEffect().getLocationChanged();
+    private void handleEvent(MachineEvent event, Integer changeId, long sequenceId, ClaimEffect e) {
+        ShopEffectUnit shopEffect = e.getShopEffect();
+        ShopDetails detailsChanged = shopEffect.getEffect().getDetailsChanged();
         String shopId = shopEffect.getShopId();
         String partyId = event.getSourceId();
-        log.info("Start shop locationChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
+        log.info("Start shop detailsChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
                 sequenceId, partyId, shopId, changeId);
-        Party party = partyRepository.findById(partyId).orElseThrow(PartyNotFoundException::new);
 
-        Shop shop = party.getShops().get(0);
+        Party party = partyRepository.findById(partyId).orElseThrow(() -> new PartyNotFoundException(partyId));
 
-        if (locationChanged.isSetUrl()) {
-            shop.setLocationUrl(locationChanged.getUrl());
-        } else {
-            throw new IllegalArgumentException("Illegal shop location " + locationChanged);
-        }
+        Shop shop = party.getShopById(shopId).orElseThrow(() -> new ShopNotFoundException(shopId));
+
+        shop.setDetailsName(detailsChanged.getName());
+        shop.setDetailsDescription(detailsChanged.getDescription());
 
         partyRepository.save(party);
 
-        log.info("End shop locationChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
+        log.info("End shop detailsChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
                 sequenceId, partyId, shopId, changeId);
     }
 }
