+++ src/main/java/com/rbkmoney/deanonimus/kafka/handler/party_management/shop/ShopLocationChangedHandler.java
@@ -1,13 +1,15 @@
-package com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.shop;
+package com.rbkmoney.deanonimus.kafka.handler.party_management.shop;
 
+import com.rbkmoney.damsel.domain.ShopLocation;
 import com.rbkmoney.damsel.payment_processing.ClaimEffect;
 import com.rbkmoney.damsel.payment_processing.PartyChange;
 import com.rbkmoney.damsel.payment_processing.ShopEffectUnit;
 import com.rbkmoney.deanonimus.db.PartyRepository;
 import com.rbkmoney.deanonimus.db.exception.PartyNotFoundException;
+import com.rbkmoney.deanonimus.db.exception.ShopNotFoundException;
 import com.rbkmoney.deanonimus.domain.Party;
 import com.rbkmoney.deanonimus.domain.Shop;
-import com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.AbstractClaimChangedHandler;
+import com.rbkmoney.deanonimus.kafka.handler.party_management.AbstractClaimChangedHandler;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,7 +22,7 @@ import java.util.List;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ShopPayoutToolChangedHandler extends AbstractClaimChangedHandler {
+public class ShopLocationChangedHandler extends AbstractClaimChangedHandler {
 
     private final PartyRepository partyRepository;
 
@@ -30,28 +32,32 @@ public class ShopPayoutToolChangedHandler extends AbstractClaimChangedHandler {
         long sequenceId = event.getEventId();
         List<ClaimEffect> claimEffects = getClaimStatus(change).getAccepted().getEffects();
         for (ClaimEffect claimEffect : claimEffects) {
-            if (claimEffect.isSetShopEffect() && claimEffect.getShopEffect().getEffect().isSetPayoutToolChanged()) {
+            if (claimEffect.isSetShopEffect() && claimEffect.getShopEffect().getEffect().isSetLocationChanged()) {
                 handleEvent(event, changeId, sequenceId, claimEffect);
             }
         }
     }
 
-    private void handleEvent(MachineEvent event, Integer changeId, long sequenceId, ClaimEffect e) {
-        ShopEffectUnit shopEffect = e.getShopEffect();
-        String payoutToolChanged = shopEffect.getEffect().getPayoutToolChanged();
+    private void handleEvent(MachineEvent event, Integer changeId, long sequenceId, ClaimEffect claimEffect) {
+        ShopEffectUnit shopEffect = claimEffect.getShopEffect();
+        ShopLocation locationChanged = shopEffect.getEffect().getLocationChanged();
         String shopId = shopEffect.getShopId();
         String partyId = event.getSourceId();
-        log.info("Start shop payoutToolChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
+        log.info("Start shop locationChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
                 sequenceId, partyId, shopId, changeId);
+        Party party = partyRepository.findById(partyId).orElseThrow(() -> new PartyNotFoundException(partyId));
 
-        Party party = partyRepository.findById(partyId).orElseThrow(PartyNotFoundException::new);
-        Shop shop = party.getShops().get(0);
+        Shop shop = party.getShopById(shopId).orElseThrow(() -> new ShopNotFoundException(shopId));
 
-        shop.setPayoutToolId(payoutToolChanged);
+        if (locationChanged.isSetUrl()) {
+            shop.setLocationUrl(locationChanged.getUrl());
+        } else {
+            throw new IllegalArgumentException("Illegal shop location " + locationChanged);
+        }
 
         partyRepository.save(party);
 
-        log.info("End shop payoutToolChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
+        log.info("End shop locationChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
                 sequenceId, partyId, shopId, changeId);
     }
 }
