+++ src/main/java/com/rbkmoney/deanonimus/kafka/handler/party_management/shop/ShopPayoutScheduleChangedHandler.java
@@ -1,13 +1,15 @@
-package com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.shop;
+package com.rbkmoney.deanonimus.kafka.handler.party_management.shop;
 
 import com.rbkmoney.damsel.payment_processing.ClaimEffect;
 import com.rbkmoney.damsel.payment_processing.PartyChange;
+import com.rbkmoney.damsel.payment_processing.ScheduleChanged;
 import com.rbkmoney.damsel.payment_processing.ShopEffectUnit;
 import com.rbkmoney.deanonimus.db.PartyRepository;
 import com.rbkmoney.deanonimus.db.exception.PartyNotFoundException;
+import com.rbkmoney.deanonimus.db.exception.ShopNotFoundException;
 import com.rbkmoney.deanonimus.domain.Party;
 import com.rbkmoney.deanonimus.domain.Shop;
-import com.rbkmoney.deanonimus.kafka.handler.party_mngmnt.AbstractClaimChangedHandler;
+import com.rbkmoney.deanonimus.kafka.handler.party_management.AbstractClaimChangedHandler;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,7 +22,7 @@ import java.util.List;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ShopCategoryChangedHandler extends AbstractClaimChangedHandler {
+public class ShopPayoutScheduleChangedHandler extends AbstractClaimChangedHandler {
 
     private final PartyRepository partyRepository;
 
@@ -30,7 +32,7 @@ public class ShopCategoryChangedHandler extends AbstractClaimChangedHandler {
         long sequenceId = event.getEventId();
         List<ClaimEffect> claimEffects = getClaimStatus(change).getAccepted().getEffects();
         for (ClaimEffect claimEffect : claimEffects) {
-            if (claimEffect.isSetShopEffect() && claimEffect.getShopEffect().getEffect().isSetCategoryChanged()) {
+            if (claimEffect.isSetShopEffect() && claimEffect.getShopEffect().getEffect().isSetPayoutScheduleChanged()) {
                 handleEvent(event, changeId, sequenceId, claimEffect);
             }
         }
@@ -38,20 +40,24 @@ public class ShopCategoryChangedHandler extends AbstractClaimChangedHandler {
 
     private void handleEvent(MachineEvent event, Integer changeId, long sequenceId, ClaimEffect e) {
         ShopEffectUnit shopEffect = e.getShopEffect();
-        int categoryId = shopEffect.getEffect().getCategoryChanged().getId();
+        ScheduleChanged payoutScheduleChanged = shopEffect.getEffect().getPayoutScheduleChanged();
         String shopId = shopEffect.getShopId();
         String partyId = event.getSourceId();
-        log.info("Start shop categoryId changed handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
+        log.info("Start shop payoutScheduleChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
                 sequenceId, partyId, shopId, changeId);
 
-        Party party = partyRepository.findById(partyId).orElseThrow(PartyNotFoundException::new);
-        Shop shop = party.getShops().get(0);
+        Party party = partyRepository.findById(partyId).orElseThrow(() -> new PartyNotFoundException(partyId));
+        Shop shop = party.getShopById(shopId).orElseThrow(() -> new ShopNotFoundException(shopId));
 
-        shop.setCategoryId(categoryId);
+        if (payoutScheduleChanged.isSetSchedule()) {
+            shop.setPayoutScheduleId(payoutScheduleChanged.getSchedule().getId());
+        } else {
+            shop.setPayoutScheduleId(null);
+        }
 
         partyRepository.save(party);
 
-        log.info("End shop categoryId changed handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
+        log.info("End shop payoutScheduleChanged handling, sequenceId={}, partyId={}, shopId={}, changeId={}",
                 sequenceId, partyId, shopId, changeId);
     }
 }
