+++ src/main/java/com/rbkmoney/deanonimus/kafka/listener/PartyManagementListener.java
@@ -6,7 +6,6 @@ import com.rbkmoney.machinegun.eventsink.SinkEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.springframework.core.annotation.Order;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.stereotype.Component;
@@ -17,7 +16,6 @@ import java.util.stream.Collectors;
 @Slf4j
 @RequiredArgsConstructor
 @Component
-@Order(value = Integer.MIN_VALUE)
 public class PartyManagementListener {
 
     private final PartyManagementHandlerService partyManagementHandlerService;
